백준 > 실버3 > 계단오르기 (2579번) - 성공 (DP 알고리즘 / Bottom-up) R
